Move all related files from ApiBundle and AdminBundle to CoreBundle
