+++ src/Sylius/Bundle/AdminBundle/Resources/config/app/sylius/sylius_mailer.yml
@@ -9,6 +9,3 @@ sylius_mailer:
         shipment_confirmation:
             subject: sylius.emails.shipment_confirmation.subject
             template: "@SyliusAdmin/Email/shipmentConfirmation.html.twig"
-        admin_password_reset:
-            subject: sylius.emails.admin_password_reset.subject
-            template: "@SyliusAdmin/Email/adminPasswordReset.html.twig"
