+++ src/Sylius/Bundle/AdminBundle/Resources/translations/messages.en.yml
@@ -6,11 +6,6 @@ sylius:
             thank_you_for_transaction: 'Thank you for a successful transaction.'
             you_can_check_its_location_with_the_tracking_code: 'You can check it''s location with the %tracking_code% tracking code.'
             your_order_with_number: 'Your order with number'
-        admin_password_reset:
-            hello: 'Hello'
-            subject: 'Administrator password reset'
-            to_reset_your_password_token: 'To reset your password - use the token below'
-            token: 'Your password reset token is'
     menu:
         admin:
             main:
