+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_resources/ResetPassword.xml
@@ -37,7 +37,7 @@
             <collectionOperation name="admin_create_reset_password_request">
                 <attribute name="method">POST</attribute>
                 <attribute name="path">/admin/reset-password-requests</attribute>
-                <attribute name="input">Sylius\Bundle\ApiBundle\Command\Admin\RequestResetPasswordEmail</attribute>
+                <attribute name="input">Sylius\Bundle\CoreBundle\Message\Admin\Account\RequestResetPasswordEmail</attribute>
                 <attribute name="output">false</attribute>
                 <attribute name="status">202</attribute>
                 <attribute name="denormalization_context">
