+++ src/Sylius/Bundle/CoreBundle/MessageHandler/Admin/Account/RequestResetPasswordEmailHandler.php
@@ -11,10 +11,10 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\CommandHandler\Admin;
+namespace Sylius\Bundle\CoreBundle\MessageHandler\Admin\Account;
 
-use Sylius\Bundle\ApiBundle\Command\Admin\RequestResetPasswordEmail;
-use Sylius\Bundle\ApiBundle\Command\Admin\SendResetPasswordEmail;
+use Sylius\Bundle\CoreBundle\Message\Admin\Account\RequestResetPasswordEmail;
+use Sylius\Bundle\CoreBundle\Message\Admin\Account\SendResetPasswordEmail;
 use Sylius\Calendar\Provider\DateTimeProviderInterface;
 use Sylius\Component\Core\Model\AdminUserInterface;
 use Sylius\Component\User\Repository\UserRepositoryInterface;
