+++ src/Sylius/Bundle/CoreBundle/Resources/config/serialization/Messages/Admin/Account/RequestResetPasswordEmail.xml
@@ -13,7 +13,7 @@
             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
             xsi:schemaLocation="http://symfony.com/schema/dic/serializer-mapping https://symfony.com/schema/dic/serializer-mapping/serializer-mapping-1.0.xsd"
 >
-    <class name="Sylius\Bundle\ApiBundle\Command\Admin\RequestResetPasswordEmail">
+    <class name="Sylius\Bundle\CoreBundle\Message\Admin\Account\RequestResetPasswordEmail">
         <attribute name="email">
             <group>admin:reset_password:create</group>
         </attribute>
