+++ src/Sylius/Bundle/CoreBundle/Resources/config/services.xml
@@ -28,6 +28,7 @@
         <import resource="services/installer.xml" />
         <import resource="services/installer_requirements.xml" />
         <import resource="services/listeners.xml" />
+        <import resource="services/message_handlers.xml" />
         <import resource="services/order_processing.xml" />
         <import resource="services/payment.xml" />
         <import resource="services/promotion.xml" />
