+++ src/Sylius/Bundle/CoreBundle/Resources/translations/messages.en.yml
@@ -1,5 +1,10 @@
 sylius:
     email:
+        admin_password_reset:
+            hello: 'Hello'
+            subject: 'Administrator password reset'
+            to_reset_your_password_token: 'To reset your password - use the token below'
+            token: 'Your password reset token is'
         contact_request:
             content: 'Content'
             message_from: 'Message from'
