+++ src/Sylius/Bundle/CoreBundle/Resources/views/Email/adminPasswordReset.html.twig
@@ -1,4 +1,4 @@
-{% extends '@SyliusAdmin/Email/layout.html.twig' %}
+{% extends '@SyliusCore/Email/layout.html.twig' %}
 
 {% block subject %}
     {{ 'sylius.email.admin_password_reset.subject'|trans({}, null, localeCode) }}
